+++ src/main/kotlin/io/gladed/watchable/ReadOnlyWatchableValue.kt
@@ -16,7 +16,7 @@
 
 package io.gladed.watchable
 
-/** A value which may not be modified externally but may still be watched for changes. */
+/** A [Watchable] value of type [T] which may not be replaced by the reference holder. */
 interface ReadOnlyWatchableValue<T> : Watchable<T, ValueChange<T>> {
     /** The current value. */
     val value: T