+++ src/test/kotlin/BindMapTest.kt
@@ -17,6 +17,7 @@
 import io.gladed.watchable.MapChange
 import io.gladed.watchable.watch
 import io.gladed.watchable.watchableMapOf
+import kotlinx.coroutines.runBlocking
 import org.junit.Assert.assertEquals
 import org.junit.Assert.fail
 import org.junit.Rule
@@ -26,25 +27,19 @@ class BindMapTest {
     @Rule @JvmField val changes = ChangeWatcherRule<MapChange<Int, String>>()
 
     @Test fun bind() {
-        runThenCancel {
+        runBlocking {
             val origin = watchableMapOf(5 to "5")
             val dest = watchableMapOf(6 to "6")
             dest.bind(origin)
-            watch(dest) { changes += it }
-            changes.expect(MapChange.Initial(mapOf(5 to "5")))
-            assertEquals(mapOf(5 to "5"), dest.get())
+            eventually { assertEquals(mapOf(5 to "5"), dest.get()) }
         }
     }
 
     @Test fun bindThenChange() {
-        runThenCancel {
+        runBlocking {
             val origin = watchableMapOf(5 to "5")
             val dest = watchableMapOf(6 to "6")
-            watch(dest) { changes += it }
             dest.bind(origin)
-            changes.expect(MapChange.Initial(mapOf(6 to "6")))
-            changes.expect(MapChange.Remove(6, "6"))
-            changes.expect(MapChange.Add(5, "5"))
             origin.use {
                 this[7] = "7"
                 this -= 5
@@ -53,18 +48,13 @@ class BindMapTest {
             origin.use {
                 put(9, "9")
             }
-            changes.expect(MapChange.Add(7, "7"))
-            changes.expect(MapChange.Remove(5, "5"))
-            changes.expect(MapChange.Replace(7, "7", "77"))
-            changes.expect(MapChange.Add(9, "9"))
-
-            assertEquals(origin.get(), dest.get())
+            eventually { assertEquals(origin.get(), dest.get()) }
         }
     }
 
     @Test fun badAdd() {
         try {
-            runThenCancel {
+            runBlocking {
                 val origin = watchableMapOf(5 to "5")
                 val dest = watchableMapOf(6 to "6")
                 dest.bind(origin)
@@ -78,7 +68,7 @@ class BindMapTest {
 
     @Test fun badRemove() {
         try {
-            runThenCancel {
+            runBlocking {
                 val origin = watchableMapOf(5 to "5")
                 val dest = watchableMapOf(6 to "6")
                 dest.bind(origin)
@@ -94,21 +84,20 @@ class BindMapTest {
     }
 
     @Test fun unbindThenChange() {
-        runThenCancel {
+        runBlocking {
             val origin = watchableMapOf(5 to "5")
             val dest = watchableMapOf(6 to "6")
             dest.bind(origin)
-            watch(dest) { changes += it }
-            changes.expect(MapChange.Initial(mapOf(5 to "5")))
             origin.use {
                 putAll(listOf(8 to "8", 7 to "7"))
             }
-            changes.expect(MapChange.Add(8, "8"), MapChange.Add(7, "7"))
+            eventually { assertEquals(mapOf(5 to "5", 8 to "8", 7 to "7"), dest.get()) }
             dest.unbind()
             origin.use {
                 remove(5)
             }
-            assertEquals(mapOf(5 to "5", 8 to "8", 7 to "7"), dest.get())
+            // Make sure no changes arrive for a short time
+            always { assertEquals(mapOf(5 to "5", 8 to "8", 7 to "7"), dest.get()) }
         }
     }
 }